Add/Delete Roles and Assign/Unassign to users

+++ Pages/Identity/Admin/Dashboard.cshtml.cs
@@ -3,13 +3,20 @@ using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.Extensions.Configuration;
 
 namespace IdentityApp.Pages.Identity.Admin
 {
     public class DashboardModel : AdminPageModel
     {
-        public DashboardModel(UserManager<IdentityUser> usrMgr) => UserManager = usrMgr;
+        public DashboardModel(UserManager<IdentityUser> userMgr,
+            IConfiguration configuration)
+        {
+            UserManager = userMgr;
+            DashboardRole = configuration["Dashboard:Role"] ?? "Dashboard";
+        }
         public UserManager<IdentityUser> UserManager { get; set; }
+        public string DashboardRole { get; set; }
         public int UsersCount { get; set; } = 0;
         public int UsersUnconfirmed { get; set; } = 0;
         public int UsersLockedout { get; set; } = 0;
@@ -31,8 +38,12 @@ namespace IdentityApp.Pages.Identity.Admin
         {
             foreach (IdentityUser existingUser in UserManager.Users.ToList())
             {
-                IdentityResult result = await UserManager.DeleteAsync(existingUser);
-                result.Process(ModelState);
+                if (emails.Contains(existingUser.Email) ||
+                    !await UserManager.IsInRoleAsync(existingUser, DashboardRole))
+                {
+                    IdentityResult result = await UserManager.DeleteAsync(existingUser);
+                    result.Process(ModelState);
+                }
             }
             foreach (string email in emails)
             {
